+++ spotdl_gui/__main__.py
@@ -1,4 +1,4 @@
-from spotdl_gui.newGui import main
+from spotdl_gui.app import main
 
 if __name__ == "__main__":
     main()
